Split lineage and completion parse trees into separate files.
